Improve `htlc_maximum_msat` in channel updates (#2299)
* Improve htlc_maximum_msat in channel updates

We previously set the `htlc_maximum_msat` inside `channel_update` to the
channel's capacity, but that didn't make any sense: we will reject htlcs
that are above the local or remote `max_htlc_value_in_flight_msat`.

We now set this value to match the lowest `max_htlc_value_in_flight_msat`
of the channel, and properly type our local value to be a millisatoshi
amount instead of a more generic UInt64.

* Set max-htlc-in-flight based on channel capacity

We introduce a new parameter to set `max-htlc-value-in-flight` based on
the channel capacity, when it provides a lower value than the existing
`max-htlc-value-in-flight-msat` static value.

* Allow disabling max-htlc-value-in-flight-msat

When opening a channel to a mobile wallet user, we may want to set our
`max-htlc-value-in-flight-msat` to something greater than the funding
amount to allow the wallet user to empty their channels using a single
HTLC per channel.

+++ eclair-core/src/main/scala/fr/acinq/eclair/channel/fsm/ChannelOpenDualFunded.scala
@@ -27,7 +27,7 @@ import fr.acinq.eclair.channel.fsm.Channel._
 import fr.acinq.eclair.channel.publish.TxPublisher.SetChannelId
 import fr.acinq.eclair.transactions.Scripts
 import fr.acinq.eclair.wire.protocol._
-import fr.acinq.eclair.{Features, RealShortChannelId, ToMilliSatoshiConversion}
+import fr.acinq.eclair.{Features, RealShortChannelId, ToMilliSatoshiConversion, UInt64}
 
 /**
  * Created by t-bast on 19/04/2022.
@@ -123,7 +123,7 @@ trait ChannelOpenDualFunded extends DualFundingHandlers with ErrorHandlers {
         commitmentFeerate = input.commitTxFeerate,
         fundingAmount = input.fundingAmount,
         dustLimit = input.localParams.dustLimit,
-        maxHtlcValueInFlightMsat = input.localParams.maxHtlcValueInFlightMsat,
+        maxHtlcValueInFlightMsat = UInt64(input.localParams.maxHtlcValueInFlightMsat.toLong),
         htlcMinimum = input.localParams.htlcMinimum,
         toSelfDelay = input.localParams.toSelfDelay,
         maxAcceptedHtlcs = input.localParams.maxAcceptedHtlcs,
@@ -165,7 +165,7 @@ trait ChannelOpenDualFunded extends DualFundingHandlers with ErrorHandlers {
             temporaryChannelId = open.temporaryChannelId,
             fundingAmount = d.init.fundingContribution_opt.getOrElse(0 sat),
             dustLimit = localParams.dustLimit,
-            maxHtlcValueInFlightMsat = localParams.maxHtlcValueInFlightMsat,
+            maxHtlcValueInFlightMsat = UInt64(localParams.maxHtlcValueInFlightMsat.toLong),
             htlcMinimum = localParams.htlcMinimum,
             minimumDepth = minimumDepth.getOrElse(0),
             toSelfDelay = localParams.toSelfDelay,
